+++ tests/test_bargraphs.py
@@ -10,7 +10,7 @@ import perun.cli as cli
 import perun.logic.store as store
 import perun.view.bars.factory as bargraphs
 
-import tests.helpers.asserts as asserts
+import tests.testing.asserts as asserts
 
 __author__ = 'Tomas Fiedor'
