+++ tests/test_detection_methods.py
@@ -10,7 +10,7 @@ import os
 import perun.logic.store as store
 from perun.utils.log import aggregate_intervals
 from perun.utils.structs import PerformanceChange
-from tests.helpers.degradation_results import (
+from tests.testing.mock_results import (
     PARAM_EXPECTED_RESULTS,
     NONPARAM_EXPECTED_RESULTS
 )
