+++ tests/test_fuzzing.py
@@ -10,7 +10,7 @@ from click.testing import CliRunner
 
 import perun.cli as cli
 
-import tests.helpers.asserts as asserts
+import tests.testing.asserts as asserts
 
 
 @pytest.mark.usefixtures('cleandir')
