+++ tests/test_regression_analysis.py
@@ -11,8 +11,8 @@ import pytest
 
 import perun.utils.exceptions as exceptions
 from perun.postprocess.regression_analysis.run import postprocess
-import tests.helpers.utils as test_utils
-from tests.helpers.utils import compare_results, generate_models_by_uid
+import tests.testing.utils as test_utils
+from tests.testing.utils import compare_results, generate_models_by_uid
 
 __author__ = 'Jiri Pavela'
