+++ tests/test_scatter_plot.py
@@ -8,7 +8,7 @@ from click.testing import CliRunner
 import perun.cli as cli
 import perun.view.scatter.factory as scatter
 
-import tests.helpers.asserts as asserts
+import tests.testing.asserts as asserts
 
 __author__ = 'Jiri Pavela'
